Some updates to the layaout opptimization project
I did these changes before the easter but forgot to commit them. I made
these changes to get content for the report. The changes are just some
new tests to check the layout efficiencies on different strings, and
some methods for generating latex code for the result so it could be
used in the report.